:recycle: move register user service to its src file
