+++ backend/src/data/providers/generate-id.ts
@@ -1,0 +1,3 @@
+export interface RandomIdGeneratorProvider {
+  generate(): string;
+}
