+++ backend/src/data/providers/index.ts
@@ -1,1 +1,2 @@
 export * from "./get-random-gift-provider";
+export * from "./generate-id";
